Use same video creation method for player and thumbnailer
